More options for PIMs
Hi all,

To allow for different resolutions of the PIMs surface model such as ZoomF=1,2,4 or 8 (similar to Malt), I added the option to set ZoomF for Pims2MNT
	- in cAppli_MPI2Mnt::cAppli_MPI2Mnt
	- and void cAppli_MPI2Mnt::DoMTD()
This allows to set ZoomF independently for the surface reconstruction (PIMs) and the interpolation (Pims2MNT).

The old default behavior would now be:
	- Pims Forest .*.JPG Referenced ZoomF=4
	-	Pims2MNT Forest DoOrtho=0 ZoomF=2
or as the default:
	- Pims Forest .*.JPG Referenced
	- Pims2MNT Forest DoOrtho=0

With the update it is now possible to set:
Pims Forest .*.JPG Referenced ZoomF=1
Pims2MNT Forest DoOrtho=0 ZoomF=1

I also added options to change the default names of DirMTD, DirOrtho, DirBasc, NameMerge
	- in cAppli_MPI2Mnt::cAppli_MPI2Mnt
such as:
Pims2MNT Forest DoOrtho=0 ZoomF=1 DirMTD=PIMs-Tmp/ DirOrtho=PIMs-Ortho/ DirBasc=PIMs-DSM/ NameMerge=DSM.xml

This is my first pull request. I hope it is ok as is and can be merged.

Best,
Thorsten