add thread control through an env variable
The number of maximum usable threads can be controlled by the environment
variable MICMAC_MAX_THREADS. If not setted, all available threads will be
used (previous default). If setted with an integer, the given number will
be considered as the max number of usable threads.

CAUTION : there is no safeguard for the following edge cases :
* given integer is higher than available thread;
* given value is not an integer.